Move backend protocols and types to a separate file from HttpClient
